+++ jsktoolbox/__init__.py
@@ -12,6 +12,6 @@ networking tools, logging infrastructure, and device-specific integrations.
 from typing import Tuple
 
 __author__ = "Jacek 'Szumak' Kotlarski"
-__version_info__: Tuple[int, int, int] = (1, 2, 9)  # Major, Minor, Patch
+__version_info__: Tuple[int, int, int] = (1, 2, 10)  # Major, Minor, Patch
 __suffix__: str = "-DEV"  # Optional suffix for pre-release versions
 __version__: str = ".".join(map(str, __version_info__)) + __suffix__